Update deploy stage in Jenkinsfile to pull latest Docker image and handle existing container

+++ Jenkinsfile
@@ -65,8 +65,10 @@ pipeline {
                     sh '''
                         ssh -o StrictHostKeyChecking=no root@159.223.163.29 "
                             docker pull ${DOCKER_IMAGE}:latest
-                            docker stop expressci-app || true
-                            docker rm expressci-app || true
+                            if [ $docker ps -aq -f name=expressci-app) ]; then
+                                docker stop expressci-app || true
+                                docker rm expressci-app || true
+                            fi
                             docker run -d --name expressci-app -p 3000:3000 ${DOCKER_IMAGE}:latest
                         "
                     '''
